Separation of CommandRunner.java and DRY'd Shop item methods
